add artist level handler
