+++ js/data/createLevelType.js
@@ -1,4 +1,4 @@
-﻿import levelType from './types/levelType';
+import levelType from './types/levelType';
 import audioList from './audioList';
 
 const createLevelType = (_levelType) => {
