+++ js/templates/start/welcomeTemplate.js
@@ -1,4 +1,4 @@
-﻿import logoTemplate from '../parts/logoTemplate';
+import logoTemplate from '../parts/logoTemplate';
 import renderScreen from '../../utils/renderScreen';
 
 const welcomeTemplate = (level) => `<section class="main main--welcome">
@@ -15,12 +15,12 @@ const welcomeTemplate = (level) => `<section class="main main--welcome">
 export default welcomeTemplate;
 
 export const welcomeHandler = function (template, state) {
-    const playButton = document.querySelector(`.main-play`);
-    playButton.addEventListener(`click`, startGame.bind(playButton, template, state));
+  const playButton = document.querySelector(`.main-play`);
+  playButton.addEventListener(`click`, startGame.bind(playButton, template, state));
 };
 
 const startGame = function (...args) {
-    const [template, state] = args;
-    renderScreen(template, Object.assign({}, state, {}));
-    this.removeEventListener(`click`, startGame.bind(this, template, state));
+  const [template, state] = args;
+  renderScreen(template, Object.assign({}, state, {}));
+  this.removeEventListener(`click`, startGame.bind(this, template, state));
 };
